+++ oversteer/gtk_handlers.py
@@ -192,6 +192,10 @@ class GtkHandlers:
     def on_launch_options_copy_clicked(self, widget):
         self.ui.copy_launch_options()
 
+    def on_rev_leds_shift_value_changed(self, widget):
+        if not self.ui.updating_rev_leds:
+            self.model.set_rev_leds_shift(widget.get_value_as_int())
+
     def on_rev_leds_test_clicked(self, widget):
         self.controller.test_rev_leds()